Beginning model refactoring
