+++ VideoImageDeltaApp/Models/PreviewType.cs
@@ -1,0 +1,11 @@
+﻿namespace VideoImageDeltaApp.Models
+{
+    public enum PreviewType
+    {
+        Video,
+        Feed,
+        Screen,
+        WatchZone,
+        Watcher
+    }
+}
