+++ VideoImageDeltaApp/Models/ffprobe.cs
