+++ VideoImageDeltaApp/PreviewBox.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Drawing;
+using System.Data;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace VideoImageDeltaApp
+{
+    public partial class PreviewBox : UserControl
+    {
+        public PreviewBox()
+        {
+            InitializeComponent();
+        }
+
+        private void PreviewBox_Click(object sender, EventArgs e)
+        {
+
+        }
+    }
+}
